+++ pyzoo/test/zoo/automl/regression/test_time_sequence_predictor.py
@@ -16,7 +16,7 @@
 import pytest
 
 from test.zoo.pipeline.utils.test_utils import ZooTestCase
-from zoo.zouwu.model.forecast.model import BaseModel
+from zoo.zouwu.model.forecast.model.abstract import BaseModel
 from zoo.automl.regression.time_sequence_predictor import TimeSequencePredictor
 import pandas as pd
 import numpy as np
