+++ pyzoo/zoo/automl/search/ray_tune_search_engine.py
@@ -28,6 +28,7 @@ from zoo.zouwu.model.forecast.model import ModelBuilder
 from zoo.zouwu.feature.identity_transformer import IdentityTransformer
 from zoo.automl.search.tune_utils import (create_searcher,
                                           create_scheduler)
+from zoo.zouwu.preprocessing.impute import LastFillImpute, FillZeroImpute
 
 SEARCH_ALG_ALLOWED = ("variant_generator", "skopt", "bayesopt")
