+++ pyzoo/zoo/zouwu/preprocessing/impute/__init__.py
@@ -13,3 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+from .impute import LastFillImpute, FillZeroImpute, TimeMergeImputor
